fix(sonar): clear new-code quality gate on fbs-gen merge
Two gate conditions failed after the in-house FlatBuffers toolchain landed:
new_reliability_rating (6 BUGs) and new_duplicated_lines_density.

Reliability — 5× S2184 (int offset arithmetic widened to a long index, a
real overflow guard) and 1× S6218:
- FbsBuilder vtable dedup/patch: widen the segment-index operands to long
  (the soffset *value* at the patch site stays int by design).
- FlatSegmentDecoder / InspectorTree peekFlatRoot: read the trailing length
  field at `segLen - 4L`; InspectorTree's fbStart becomes long to match its
  bounds-checked twin in FlatSegmentDecoder.
- TypeRegistry.StructLayout: @SuppressWarnings("java:S6218"); codegen-internal
  carrier, fieldOffsets never compared (same precedent as UnpackSchedule).

Duplication — exclude the two build-only code generators (fbs-gen, proto-gen)
from CPD via sonar.cpd.exclusions. Their lexers share tokenizer shape by
design; smell and bug analysis on them is kept.

+++ core/src/main/java/io/github/dfa1/vortex/fbsrt/FbsBuilder.java
@@ -328,7 +328,7 @@ public final class FbsBuilder {
                 continue;
             }
             for (int k = SIZEOF_SHORT; k < len; k += SIZEOF_SHORT) {
-                if (buf.get(LE_SHORT, vt1 + k) != buf.get(LE_SHORT, vt2 + k)) {
+                if (buf.get(LE_SHORT, (long) vt1 + k) != buf.get(LE_SHORT, (long) vt2 + k)) {
                     continue outer;
                 }
             }
@@ -344,7 +344,7 @@ public final class FbsBuilder {
                 vtables = Arrays.copyOf(vtables, numVtables * 2);
             }
             vtables[numVtables++] = offset();
-            buf.set(LE_INT, cap - vtableloc, offset() - vtableloc);
+            buf.set(LE_INT, (long) cap - vtableloc, offset() - vtableloc);
         }
         return vtableloc;
     }
